Update Reactive Forces
In this lab, we are introduced to reactive forces (Newton's 3rd Law), static friction, and kinetic friction, for an object on a ramp. The object on the ramp already has forces acting on it, these forces are the tangent force, parallel to the ramp, and the normal force, perpendicular to the ramp. The frictional forces we would apply to the object are the reactive forces that we can visualize to be of the opposite but equal magnitude of the tangent force.

In our lab, we determine that our acting force is the sum of our perpendicular acting force (Normal force) and parallel acting force (tangent force). We then compute the normal force and the tangent force by using two different equations. Normal force = - (normal vector * acting force) * normal vector, Tangent Force = - ( tangent vector * acting force) * tangent vector. From here, we are able to determine our frictional forces. If our reactive force is less than the static coefficient multiplied by the normal force, we are given static friction. If our reactive force is greater than or equal to the static coefficient multiplied by the normal force, we are given kinetic friction.

At the start of the lab we calculated an object in motion along a ramp of 35 degrees, given static friction coefficient and kinetic friction coefficient of 0. We run this experiment with a gravitational force of -9.81. This object also had a force applied to it at 0.22s to 0.38s of 0.001 N. Here we see that the position of the object is moving downward on the ramp with no friction, if the velocity graph was depicted it would show a line starting at position 0 moving in a consistently negative direction at each time step. On our first acceleration graph we see that the acceleration remains constant along a horizontal line with the small change at 0.22s to 0.38s where the force of 0.001 N is applied. We have created these graphs to show the difference in position, velocity, and acceleration of the object moving along the ramp without reactive forces applied.

Static friction is the force applied to an object of equal magnitude to the acting force. When static friction is applied, the object will remain static or at rest because the reactive force fails to exceed the racting forces on the object. In our next experiment, we add a reactive force of static friction to the object. Here we can see the net acceleration, given the angle of the ramp and the static friction coefficient. Our graph shows us that when the angle of the ramp increases, our static coefficient also increases. Where you see solid blue on our acceleration graph, this means the object is at rest. For example, at a 0 degree ramp, our static friction coefficient needed to keep the object at rest is also 0 N.. Another example can be shown when our ramp is at 30 degrees, our static friction coefficient is approximately 0.6 N in order for our object to remain at rest on an angled ramp.

We performed another experiment for static friction using the parameters of a ramp angle of 90 degrees and static friction coefficient maximum of 60 N. In this experiment, we were trying to test different values of static friction coefficient to get an object on a 90 degree ramp to remain at rest. On this graph, you will notice that at an extreme force of 60 N, our object is not static. With further experimentation, we eventually see that there is no value of force strong enough to keep our object at rest on a 90 degree incline ramp.

When static friction is no longer able to match the acting force we have a force breaking point. This is the moment when static friction becomes kinetic friction. Kinetic friction is equal to the force that is applied to an object when it is greater than static friction allowing the object to move. For example, if a force is applied to our object stronger than static friction our object will begin to slide down the ramp, kinetic friction is the force that is reacting on the object in the opposite direction of the object's movement. To calculate kinetic friction force, we would multiply the kinetic friction coefficient to the normal force which is then multiplied by the tangent vector to the surface.